Rm TraceBuilder.set_ast_context, read from TracingState, merge builder->state.py
